Added back protocol-deployments package - for bundling all deployments into one npm package (#139)
Added back the `protocol-deployments` package, but this time it just contains the wagmi config and typescript to publish to npm.  This will also eventually contain abis/addresses for the new MINT contract as well as things like the universal minter

The idea is that this new package package imports the addresses and abis from other deployment packages and bundles and publishes them.  This is the package that is published for deployed addresses and abis, and should be the one imported externally now.

This was done to share wagmi config across projects, and create a single easy to use discoverable npm package.

Other advantages:
* individual deployment projects dont need wagmi anymore
* when importing `protocol-deployments` - it is a pure js package, there is no forge/foundry/solidity code or dependencies.
* We can have a clean readme just around importing and using this library which is discoverable from npmjs.com