Fork Tests code gen for each chain (#361)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

As it exists, fork tests run through a loop through each chain then run the test for the chain.  It's challenging to determine which chain's fork test is broken, and it takes each fork tests a while to run as it has to go through all chains before getting a result.

This pr solves this issue through code generation: it generates a test for each chain for each fork test, allowing the test for each fork to run in parallel, and enabling the error to be shown for each chain.

We can still use FORK_TEST_CHAINS to filter out which fork test to run. Otherwise, it runs the fork tests for all chains.

Also remove MintsTest that checks preminting with MINTs - thats already tested in the upgrades test. 

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## Does this change the ABI/API?

- [ ] This changes the ABI/API

<!-- If so, please describe how and what potential impact this may have -->

## What tests did you add/modify to account for these changes

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New module / feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] i added a changeset to account for this change

## Reviewer Checklist:

- [ ] My review includes a symposis of the changes and potential issues
- [ ] The code style is enforced
- [ ] There are no risky / concerning changes / additions to the PR