protocol-sdk and protocol-deployments docs (#412)
* remove erc20 minter abi from checked-in code (#418)

When I run `yarn build` in the root of the project, `packages/creator-subgraph/abis/ERC20Minter.json` gets modified and then git asks me to commit that file, even though my code has not touched anything in creator-subgraph.

Leaving it `gitignored` like the rest of the abis has the behavior go away, and also, since it is just code generated from another part of the project, we shouldn't need to check it in.

* fix build step

* update site deployment task

* pull base path from env