New SDK methods: getToken, getTokensOfContract (#465)
New SDK methods:

`getToken`
The Collector Client can be used to get token 1155s, 721s, and premints with
a function `getToken()`. The type of token get is determined by the `mintType` parameter, which
can be set to either `1155`, `721`, or `premint`. Depending on the `mintType` additional parameters
must be passed.  The `getToken()` function returns both information about the token, in the `token` object, and a function
`prepareMint` which takes a quantity to mint and returns the transaction parameters to mint the token, costs to mint the token, and any necessary
erc20 approvals that must be executed before minting the token.

`getTokensOfContract`
**Both onchain** and **premint (gaslessly created)** tokens of a Zora 1155 contract can be retrieved using the the Collector Client function 
`getTokensOfContract()` and passing a token contract address. The function returns a `tokens` array with an item for each token of the contract.
Each item in `tokens` contains information the token in the `token` object, and has a function
`prepareMint`.  The `prepareMint` function takes a quantity to mint and returns the transaction parameters to mint the token, costs to mint the token, and any necessary
erc20 approvals that must be executed before minting the token.