SDK add ability to mint right after creating (#672)
Sometimes it is desired to be able to mint right after creating.  For this case, we shouldn't have to wait for the subgraph to index the token to get the minter settings.

This update to the SDK enables minting right after creating, by having the `create` functions return a `prepareMint` function