SDK - for creating token on existing contract, avoid multiple rpc reads and just use subgraph to get the contract info. (#675)
When creating tokens on existing contracts using the SDK, we were doing three rpc reads, which can slow things down.

This PR updates to just do a single subgraph read to get the contract version, next token id and mint fee.

It also enables retrying in case the subgraph doesnt return anything, for the case a token is to be created on the contract using the sdk  right after the contract is created, and the subgraph hasn't indexed yet