Fix deployment generation caching: * protocol-deployments-gen output goes into its own folder, which can be included with cached output. * protocol-deployments can copy that generated file to itself (#712)
the output of protocol-deployments-gen is not currently cached, because it gets put into another package (`protocol-deployments`) - this results sometimes  in ci that `protocol-deploymnets` is missing generated wagmi config if it relies on a cached `protocol-deployments-gen` step.

This PR updates `protocol-deployments-gen` to generate the file in its own package so that it can be included in cached output.  `protocol-deployments` can now copy that file over for its build step