Fix secondary market filtering issue with SDK Closes public:#452 (#826)
## Description

This change modifies the filtering logic for sales strategies in the `getTargetStrategy` function. It now includes strategies where either `saleActive` or `secondaryMarketActive` is true, expanding the criteria for valid strategies.

## Motivation and Context

The update allows for a more inclusive selection of valid sales strategies, potentially accommodating scenarios where secondary market activity is relevant even if the primary sale is not active.

## Does this change the ABI/API?

- [ ] This changes the ABI/API

## What tests did you add/modify to account for these changes

Existing tests should be updated to include cases where `secondaryMarketActive` is true but `saleActive` is false, ensuring the new filtering logic works as intended.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New module / feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I added a changeset to account for this change

## Reviewer Checklist:

- [ ] My review includes a symposis of the changes and potential issues
- [ ] The code style is enforced
- [ ] There are no risky / concerning changes / additions to the PR