coin factory - added deployWithHook (#1118)
Coin factory - added ability deploy with hook, enabling to deploy and buy currency

We need the ability to buy initial supply of a coin, when the coin is priced in Zora. Normally this takes a few transaction: swapping for the currency, approving the currency to be used by the factory contract, then calling the deploy function on the factory.

This PR adds a method to the factory, `deployWithHook`, that lets this happen in an atomic step:

* the factory will create a coin
* the factory will then call a hook with calldata passed to the function.

There is an initial hook created - `BuySupplyWithEthUsingV3Hook` - that buys supply with eth

* Added coins scripts to deploy the latest factory, as well as the hooks, and update the addresses.jsons.
* Updated protocol-deployments to include the new deploy hooks for coins.
* Have the coins sdk import addresses and abis from protocol-deployments, as that will significantly reduce its dependency tree, and also have it follow the rest of the packages.
* ensure you can only upgrade to a new Coiin factory with the same contract name 
* protocol-deployments-gen now pulls uniswap abis from shared-contracts, which now publishes its abis, instead of it getting it thru erc20z
* adding helper methods to protocol-deployments to help with encoding of hook data.