fix: resolve new fee not used when migrating liquidity (#1402)
Theres a bug that when you migrate liquidity to a new hook that has a changed fee, the new fee is not used.

This fix makes sure that the fee is updated when you migrate liquidity.   For existing hooks that migrate liquidity, they need to migrate twice to the new hook to pick up the new fee.