remove indexed label to prevent indexer parsing error (#1414)
# Remove indexed attribute from label parameter in AllocationsSet event

## Description

This PR removes the `indexed` attribute from the `label` parameter in the `AllocationsSet` event within the `IZoraIncentiveClaim` interface.

We removed this sometime last week to resolve a indexer bug @ninazo ran into, we have not established a pattern on the backend yet for consuming the hashed string that is emitted, so the simplest solution was to not index the label.

I think think is fine as it was more of a nice to have to index the label.

## Motivation and Context


## Does this change the ABI/API?

- [x] This changes the ABI/API

This change modifies the event signature in the ABI, which could affect services that are filtering for this event using the indexed parameters.

## What tests did you add/modify to account for these changes

Existing tests were verified to ensure they still pass with this change. No additional tests were needed as this is a simple modification to an event definition.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I added a changeset to account for this change

## Reviewer Checklist:

- [ ] My review includes a symposis of the changes and potential issues
- [ ] The code style is enforced
- [ ] There are no risky / concerning changes / additions to the PR